# INTPYTHON-617 - Improve DictField to_python performance