--- a/mongoengine/base/datastructures.py
+++ b/mongoengine/base/datastructures.py
@@ -472,3 +472,60 @@ class LazyReference(DBRef):
 
     def __repr__(self):
         return f"<LazyReference({self.document_type}, {self.pk!r})>"
+
+
+class RawDict:
+    def __init__(self, data, deserialize_method):
+        self._data = data
+        self.deserialize_method = deserialize_method
+
+    def deserialize(self):
+        return self.deserialize_method(self._data)
+
+    def __setitem__(self, key, item):
+        self._data[key] = item
+
+    def __getitem__(self, key):
+        return self._data[key]
+
+    def __repr__(self):
+        return repr(self._data)
+
+    def __len__(self):
+        return len(self._data)
+
+    def __delitem__(self, key):
+        del self._data[key]
+
+    def clear(self):
+        return self._data.clear()
+
+    def copy(self):
+        return self._data.copy()
+
+    def has_key(self, k):
+        return k in self._data
+
+    def update(self, *args, **kwargs):
+        return self._data.update(*args, **kwargs)
+
+    def keys(self):
+        return self._data.keys()
+
+    def values(self):
+        return self._data.values()
+
+    def items(self):
+        return self._data.items()
+
+    def pop(self, *args):
+        return self._data.pop(*args)
+
+    def __cmp__(self, dict_):
+        return self.__cmp__(self._data, dict_)
+
+    def __contains__(self, item):
+        return item in self._data
+
+    def __iter__(self):
+        return iter(self._data)