@NoahStapp 2025-08-15
> Does your benchmark still apply? You seem to have changed the code substantially, and I would assume that the `to_python` caching you do at the top level not to be that impactful (but maybe it is). If this is really the case, then it may make sense to move your change down into the base class so this applies to lists too.

What do you by mean "still apply"? The benchmarking script was run against the current master branch as of June with and then without the change to `DictField.to_python`.