@NoahStapp 2025-08-19
> > > Does your benchmark still apply? You seem to have changed the code substantially, and I would assume that the `to_python` caching you do at the top level not to be that impactful (but maybe it is). If this is really the case, then it may make sense to move your change down into the base class so this applies to lists too.
> > 
> > 
> > What do you by mean "still apply"? The benchmarking script was run against the current master branch as of June with and then without the change to `DictField.to_python`.
> 
> I wrote what I meant _right after_
> 
> > You seem to have changed the (your) code substantially
> 
> It's not obvious what commit you ran your benchmarking from, so I was just confirming. The rest of my comment applies. Have you profiled this to see what code is the bottleneck that you're removing?

Here's the snakeviz visualization of the profiling for master versus this change:

#### Master
<img width="1208" height="640" alt="Screenshot 2025-08-19 at 6 35 35 AM" src="https://github.com/user-attachments/assets/94bcf929-fcba-4297-a592-c2dc54f753a7" />
<img width="1674" height="38" alt="Screenshot 2025-08-19 at 6 40 18 AM" src="https://github.com/user-attachments/assets/8bef056d-4a04-4dd1-9bc4-68a16f38f8bf" />

#### This PR
The call graph shown is different due to nearly all of the calls from the root being PyMongo driver background operations 
<img width="1211" height="576" alt="Screenshot 2025-08-19 at 6 38 43 AM" src="https://github.com/user-attachments/assets/d4f99eaf-9e28-498a-bc11-affaa4e346b6" />
<img width="1677" height="38" alt="Screenshot 2025-08-19 at 6 38 55 AM" src="https://github.com/user-attachments/assets/74f9344a-4590-4683-9e13-7726d3b9ec45" />

The core difference shown is that in the current behavior, `DictField.to_python` makes 6,000,400 recursive calls to itself even though all of its data is primitive Python types and not other `Field` objects. 

In comparison, with this change, `DictField.to_python` makes 10 calls without any recursion.