@bagerard 2026-08-20
Hi, first of all I apologize for the delay and lack of reactivity. 
Spent a bit of time on this, first writing additional tests cases (added on this branch), and then coming up with an alternative implementation https://github.com/MongoEngine/mongoengine/pull/2924 which I believe is less risky (and perf gains are similar, it is greatly inspired by your work, the general idea is the same, it's just tackling this slightly differently). 

Note that I messed up your branch and ended up rebasing your branch on current master before adding test cases; sorry for this so pay attention if you start working on this again.

If you have a chance, please provide feedback on the other MR (#2924)